% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkthis.R
\name{chunkthis}
\alias{chunkthis}
\title{Wrap the copied code with r markdown code-chunk fence}
\description{
Simple copy a code and then go to Addins and click
Chunk the code. Also you can add a shortcut for
this Addin like any other Rstudio Addins.
}
